added tag mutations

+++ packages/server/src/modules/tag/tag.service.ts
@@ -33,18 +33,29 @@ export class TagService {
         });
     }
 
-    public async addTag(
+    public async createTag(
         user: User,
-        { label, icon }: AddTagInput
+        { label, icon, description, color, taggedItem }: AddTagInput
     ): Promise<Tag | null> {
         const tag = this.tagRepo.create({
             user: { id: user.id },
             label,
-            taggedItems: [],
+            description,
+            color,
             icon: icon ?? null,
         });
+
         const ability = this.caslAbilityFactory.createForUser(user);
         if (ability.can(EntityAction.CREATE, tag)) {
+            const foundTaggedItem = taggedItem
+                ? await this.taggableRepo.findOne(taggedItem?.id)
+                : undefined;
+            if (foundTaggedItem) {
+                await this.taggableRepo.save({
+                    ...foundTaggedItem,
+                    tags: foundTaggedItem.tags.concat(tag),
+                });
+            }
             return this.tagRepo.save(tag);
         }
         return null;
@@ -56,7 +67,37 @@ export class TagService {
     ): Promise<Tag> {
         const tag = await this.tagRepo.findOneOrFail(updateInput.id);
         tag.label = updateInput.label ?? tag.label;
+        tag.description = updateInput.description ?? tag.description;
+        tag.color = updateInput.color ?? tag.color;
         tag.icon = updateInput.icon !== undefined ? updateInput.icon : tag.icon;
+
+        if (updateInput.taggables != null) {
+            const taggables = await this.taggableRepo.findByIds(
+                updateInput.taggables.map(({ id }) => id)
+            );
+
+            // for (const taggable of taggables) {
+            //     this.taggableRepo.save({
+            //         ...taggable,
+            //         tags: taggable.tags.concat({id: tag.id})
+            //     })
+            // }
+
+            tag.taggedItems = taggables;
+        }
+
+        if (updateInput.applyTaggables?.length) {
+            for (const taggable of updateInput.applyTaggables) {
+                await this.applyTag(user, tag.id, taggable.id);
+            }
+        }
+
+        if (updateInput.removeTaggables?.length) {
+            for (const taggable of updateInput.removeTaggables) {
+                await this.removeTag(user, tag.id, taggable.id);
+            }
+        }
+
         const ability = this.caslAbilityFactory.createForUser(user);
         if (ability.can(EntityAction.UPDATE, tag)) {
             await this.tagRepo.save(tag);
@@ -66,8 +107,8 @@ export class TagService {
 
     public async applyTag(
         user: User,
-        tagId: string,
-        taggableId: string
+        tagId: number,
+        taggableId: number
     ): Promise<Tag | null> {
         const tag = await this.tagRepo.findOneOrFail(tagId);
         const taggable = await this.taggableRepo.findOneOrFail(taggableId);
@@ -83,23 +124,34 @@ export class TagService {
     }
 
     public async removeTag(
+        user: User,
         tagId: number,
         taggableId: number
     ): Promise<Tag | null> {
         try {
             const tag = await this.tagRepo.findOneOrFail(tagId);
+
             tag.taggedItems = tag.taggedItems.filter(
                 (taggable) => taggable.id !== taggableId
             );
-            return await this.tagRepo.save(tag);
+
+            const ability = this.caslAbilityFactory.createForUser(user);
+            if (ability.can(EntityAction.UPDATE, tag)) {
+                return await this.tagRepo.save(tag);
+            }
+            return null;
         } catch (err: unknown) {
             console.error(err);
             return null;
         }
     }
 
-    public async deleteTag(id: number): Promise<void> {
+    public async deleteTag(user: User, id: number): Promise<void> {
         const tag = await this.tagRepo.findOneOrFail(id);
-        await this.tagRepo.remove(tag);
+
+        const ability = this.caslAbilityFactory.createForUser(user);
+        if (ability.can(EntityAction.DELETE, tag)) {
+            await this.tagRepo.remove(tag);
+        }
     }
 }
